# Conflicts:
#	dev/Milestone3/Superly.db

+++ dev/Milestone1/src/Domain/Business/Controllers/EmployeeController.java
@@ -257,6 +257,7 @@ public class EmployeeController {
 
     //DELETE
     public void removeEmployee(String id) throws Exception {
+        if (!new ShiftController().getEmployeeShiftsBetween(id, LocalDate.now(), LocalDate.now().plusDays(30)).isEmpty()) throw new Exception("Employee assigned");
         employeeDataMapper.delete(id);
     }
